feat: remove check duplicate function
